+++ src/main/java/com/ironknee/Knee2KneelSpring/authentication/JwtUtil.java
@@ -1,10 +1,11 @@
-package com.ironknee.Knee2KneelSpring.security;
+package com.ironknee.Knee2KneelSpring.authentication;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import org.springframework.util.StringUtils;
 
 import java.util.Date;
 
@@ -40,6 +41,16 @@ public class JwtUtil {
 
     private Boolean isTokenExpired(String token) {
         return extractClaims(token).getExpiration().before(new Date());
+    }
+
+    public Claims extractTokenValue(String token) {
+        String tokenValue;
+
+        if(StringUtils.hasText(token) && token.startsWith("Bearer "))
+            tokenValue = token.substring(7);
+        else throw new NullPointerException("Token not found");
+
+        return Jwts.parserBuilder().setSigningKey(secret).build().parseClaimsJws(tokenValue).getBody();
     }
 
 }
